+++ src/triangle.rs
@@ -47,6 +47,11 @@ pub fn triangle(command_params: CommandParams, mut sketch_params: SketchParams) 
         error!("No genomes/sketches found.");
         std::process::exit(1)
     }
+
+    if ref_sketches.len() > 500 && !command_params.sparse {
+        warn!("> 500 genomes detected. The output matrix will be large. Consider using -E or --sparse for a tsv output instead.");
+    }
+
     let kmer_to_sketch = screen::kmer_to_sketch_from_refs(&ref_sketches);
     let counter: Mutex<usize> = Mutex::new(0);
     let first: Mutex<bool> = Mutex::new(true);